+++ app/Http/Controllers/AnggaranController.php
@@ -36,6 +36,7 @@ class AnggaranController extends Controller
             'message' => 'Status berhasil diubah!'
         ]);
     }
+
     public function index(Request $request)
     {
         if (!$request->tahun || !$request->jenis) {
@@ -56,12 +57,6 @@ class AnggaranController extends Controller
             $anggaran = $anggaranQuery->whereHas('detail_norekening', function ($data) {
                 $data->where('jenis_norekening_id', 6);
             })->latest()->paginate(10);
-        } elseif ($request->jenis == "laporan") {
-            $detail_norekening = Detail_Norekening::all();
-            $data = $this->laporan($request);
-            return view('anggaran.laporan', compact('detail_norekening', 'data'));
-        } elseif ($request->jenis == "grafik") {
-            return view('anggaran.grafik');
         } else {
             return redirect('anggaran?jenis=pendapatan&tahun=' . date('Y'));
         }
@@ -71,104 +66,20 @@ class AnggaranController extends Controller
         return view('anggaran.index', compact('anggaran'));
     }
 
-    public function laporan_apbdes(Request $request)
-    {
-        $detail_norekening = Detail_Norekening::all();
-        $desa = Desa::find(1);
-        $data = $this->laporan($request);
-
-        if (!$request->tahun || !$request->jenis) {
-            return redirect('laporan-apbdes?jenis=laporan&tahun=' . date('Y'));
-        }
-
-        if ($request->jenis == "laporan") {
-            return view('anggaran.laporan-apbdes', compact('desa', 'detail_norekening', 'data'));
-        } elseif ($request->jenis == "grafik") {
-            return view('anggaran.grafik-apbdes-umum', compact('desa', 'detail_norekening', 'data'));
-        } else {
-            return redirect('laporan-apbdes?jenis=laporan&tahun=' . date('Y'));
-        }
-    }
-
-    private function laporan($request)
-    {
-        $data['pendapatan_anggaran'] = 0;
-        $data['belanja_anggaran'] = 0;
-        $data['pembiayaan_anggaran'] = 0;
-
-        foreach (Anggaran::whereTahun($request->tahun)->whereHas('detail_norekening', function ($jenis) {
-            $jenis->where('jenis_norekening_id', 4);
-        })->get() as $item) {
-            $data['pendapatan_anggaran'] += $item->nilai_anggaran;
-        }
-
-        foreach (Anggaran::whereTahun($request->tahun)->whereHas('detail_norekening', function ($jenis) {
-            $jenis->where('jenis_norekening_id', 5);
-        })->get() as $item) {
-            $data['belanja_anggaran'] += $item->nilai_anggaran;
-        }
-
-        foreach (Anggaran::whereTahun($request->tahun)->whereHas('detail_norekening', function ($jenis) {
-            $jenis->where('kelompok_norekening_id', 61);
-        })->get() as $item) {
-            $data['penerimaan_biaya_anggaran'] += $item->nilai_anggaran;
-        }
-
-        foreach (Anggaran::whereTahun($request->tahun)->whereHas('detail_norekening', function ($jenis) {
-            $jenis->where('kelompok_norekening_id', 62);
-        })->get() as $item) {
-            $data['pengeluaran_biaya_anggaran'] += $item->nilai_anggaran;
-        }
-
-        return $data;
-    }
-
     public function cart(Request $request)
     {
-        $pendapatan_anggaran = 0;
-        $belanja_anggaran = 0;
-        $pembiayaan_anggaran = 0;
         $tahun = $request->tahun ? $request->tahun : date('Y');
         $rincian = [];
-
-        foreach (Anggaran::whereTahun($tahun)->whereHas('detail_norekening', function ($jenis) {
-            $jenis->where('jenis_norekening_id', 4);
-        })->get() as $item) {
-            $pendapatan_anggaran += $item->nilai_anggaran;
-        }
-
-        foreach (Anggaran::whereTahun($tahun)->whereHas('detail_norekening', function ($jenis) {
-            $jenis->where('jenis_norekening_id', 5);
-        })->get() as $item) {
-            $belanja_anggaran += $item->nilai_anggaran;
-        }
-
-        foreach (Anggaran::whereTahun($tahun)->whereHas('detail_norekening', function ($jenis) {
-            $jenis->where('jenis_norekening_id', 6);
-        })->get() as $item) {
-            $pembiayaan_anggaran += $item->nilai_anggaran;
-        }
 
         foreach (Anggaran::whereTahun($tahun)->get()->groupBy('detail_norekening_id') as $value) {
             $anggaran = 0;
             foreach ($value as $item) {
                 $anggaran += $item->nilai_anggaran;
             }
-            $rincian[] = $this->cart_rincian($value[0]->detail_norekening->jenis_norekening_id, $anggaran, $value[0]->detail_norekening->nama ? $value[0]->detail_norekening->nama : $value[0]->detail_norekening->kelompok_jenis_norekening->nama);
+            $rincian[] = $this->cart_rincian($value[0]->detail_norekening->jenis_norekening_id, $anggaran, $value[0]->detail_norekening->nama ?: $value[0]->detail_norekening->kelompok_jenis_norekening->nama);
         }
 
-        return response()->json([
-            'pendapatan' => [
-                'uang' => 'Rp. ' . substr(number_format($pendapatan_anggaran, 2, ',', '.'), 0, -3),
-            ],
-            'belanja' => [
-                'uang' => 'Rp. ' . substr(number_format($belanja_anggaran, 2, ',', '.'), 0, -3),
-            ],
-            'pembiayaan' => [
-                'uang' => 'Rp. ' . substr(number_format($pembiayaan_anggaran, 2, ',', '.'), 0, -3),
-            ],
-            'detail' => $rincian
-        ]);
+        return response()->json(['detail' => $rincian]);
     }
 
     private function cart_rincian($jenis, $anggaran, $rincian)
@@ -187,38 +98,29 @@ class AnggaranController extends Controller
     }
 
     public function store(Request $request)
-    {
-        $data = $request->validate([
-            'tahun' => ['required', 'numeric', 'min:1900'],
-            'jenis_norekening' => ['required'],
-            'detail_norekening_id' => ['required'],
-            'nilai_anggaran' => ['required', 'numeric', 'min:0'],
-            'keterangan_lainnya' => ['nullable']
-        ], [
-            'detail_norekening_id.required' => 'detail jenis anggaran wajib diisi'
-        ]);
+{
+    $data = $request->validate([
+        'tahun' => ['required', 'numeric', 'min:1900'],
+        'jenis_norekening' => ['required', 'exists:jenis_norekening,id'], // Pastikan ini sesuai
+        'detail_norekening_id' => ['required', 'exists:detail_norekening,id'],
+        'nilai_anggaran' => ['required', 'numeric', 'min:0'],
+        'keterangan_lainnya' => ['nullable']
+    ], [
+        'tahun.required' => 'Tahun wajib diisi',
+        'nilai_anggaran.min' => 'Nilai anggaran tidak boleh kurang dari 0',
+        'jenis_norekening.required' => 'Jenis norekening wajib diisi',
+        'detail_norekening_id.required' => 'Detail jenis anggaran wajib diisi'
+    ]);
 
-        $jenis = '';
-        if ($request->jenis_norekening == 4) {
-            $jenis = 'pendapatan';
-        } elseif ($request->jenis_norekening == 5) {
-            $jenis = 'belanja';
-        } elseif ($request->jenis_norekening == 6) {
-            $jenis = 'pembiayaan';
-        }
+    Anggaran::create($data);
+    return redirect('/anggaran?jenis=' . $request->jenis_norekening . "&tahun=" . $request->tahun)
+           ->with('success', 'Anggaran APBDes berhasil ditambahkan');
+}
 
-        Anggaran::create($data);
-        return redirect('/anggaran?jenis=' . $jenis . "&tahun=" . $request->tahun)->with('success', 'Anggaran APBDes berhasil ditambahkan');
-    }
 
     public function show($id)
     {
         return response()->json(Detail_Norekening::where('jenis_norekening_id', $id)->get());
-    }
-
-    public function Kelompok_Norekening(Kelompok_Norekening $Kelompok_Norekening)
-    {
-        return response()->json($Kelompok_Norekening);
     }
 
     public function edit(Anggaran $anggaran)
@@ -236,7 +138,7 @@ class AnggaranController extends Controller
             'nilai_anggaran' => ['required', 'numeric', 'min:0'],
             'keterangan_lainnya' => ['nullable']
         ], [
-            'detail_norekening_id.required' => 'detail jenis anggaran wajib diisi'
+            'detail_norekening_id.required' => 'Detail jenis anggaran wajib diisi'
         ]);
 
         $anggaran->update($data);
@@ -249,5 +151,3 @@ class AnggaranController extends Controller
         return redirect()->route('anggaran.index')->with('success', 'Anggaran APBDes berhasil dihapus');
     }
 }
-
-